==> 03_Fig3.python.ipynb <==
from svgutils.compose import *
from svgutils.compose import Figure
from pathlib import Path
from cairosvg import svg2png

FIGURE = "Figure_3"

circle="""<?xml version="1.0" encoding="UTF-8" standalone="no"?>
<svg xmlns="http://www.w3.org/2000/svg" width="50" height="50">
<circle cx="25" cy="25" r="20" fill="#fff" stroke="#000" stroke-width="1" fill-opacity="0.0"/>
</svg>"""

outfile = Path("assets/figures/Figure_3/doodles/circle.svg")
outfile.parent.mkdir(parents=True, exist_ok=True)
with open(outfile, "w") as f:
    f.write(circle)

ellipse ="""<?xml version="1.0" encoding="UTF-8" standalone="no"?>
<svg xmlns="http://www.w3.org/2000/svg" width="50" height="120">
  <ellipse cx="25" cy="60" rx="10" ry="55" fill="#fff" stroke="#000" stroke-width="1" fill-opacity="0.0"/>
</svg>"""

outfile = Path(f"assets/figures/{FIGURE}/doodles/ellipse.svg")
outfile.parent.mkdir(parents=True, exist_ok=True)
with open(outfile, "w") as f:
    f.write(ellipse)

final_figure = Figure("900", "900",
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/a.svg").scale(1),
                          Text("(a)", 0, -25, size=12, weight='bold'),
                          Text("S. spinosa/pogona", 100, 90+25, size=10),
                          Text("S. erythrea", 215, 110+45, size=10),
                      ).move(25, 25+25),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/doodles/ellipse.svg")
                      ).move(82, 80+25).rotate(-5),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/doodles/circle.svg").scale(0.6)
                      ).move(210, 145+42),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/b.svg").scale(0.7),
                          Text("(b)", 0, 0, size=12, weight='bold'),
                          Text("P1", 110, 120, size=12, weight='bold'),
                          Text("P2", 170, 180, size=12, weight='bold'),
                          Text("P5", 235, 248, size=12, weight='bold'),
                          Text("P7", 280, 295, size=12, weight='bold'),
                          Text("P8", 325, 335, size=12, weight='bold'),
                      ).move(360, 25),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/c.svg").scale(1),
                          Text("(c)", 0, 0, size=12, weight='bold'),
                      ).move(25, 425),
                     )
final_figure

final_figure.save(f"assets/figures/{FIGURE}/{FIGURE}.svg")

svg2png(url=f"assets/figures/{FIGURE}/{FIGURE}.svg", write_to=f"assets/figures/{FIGURE}/{FIGURE}.png")
