==> 07_Fig4.python.ipynb <==
from svgutils.compose import *
from svgutils.compose import Figure
from pathlib import Path
from cairosvg import svg2png
import json

FIGURE = "Figure_4"

labels = {}
targets = ['201888', '215691', '220008', '203048']

with open(f"assets/figures/{FIGURE}/{FIGURE}b.svg", "r") as f:
    xml_data = f.read().split(">")
    for target in targets:
        for num, line in enumerate(xml_data):
            if f'BiG-FAM_id: {target}' in line:
                coord = [i for i in line.split() if 'transform' in i][0].strip('transform="translate').strip("(").strip(")").split(",")
                coord = [float(i) for i in coord]
                labels[target] = coord
labels

final_figure = Figure("1200", "1200",
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}a.svg").scale(0.7).move(0, -20),
                          Text("(a)", 0, 0, size=12, weight='bold'),
                      ).move(0, 25),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}b.svg").scale(1),
                          Text("(b)", 0, 0, size=12, weight='bold'),
                          Text(f"GCF_{targets[0]}", labels[targets[0]][0]+40, labels[targets[0]][1]+5, size=6, weight='normal'),
                          Text(f"GCF_{targets[1]}", labels[targets[1]][0]+30, labels[targets[1]][1]+20, size=6, weight='normal'),
                          Text(f"GCF_{targets[2]}", labels[targets[2]][0]+40, labels[targets[2]][1]+8, size=6, weight='normal'),
                          Text(f"GCF_{targets[3]}", labels[targets[3]][0]+20, labels[targets[3]][1]+5, size=6, weight='normal'),
                      ).move(550, 25),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}c.svg").scale(0.65).move(-20, 0),
                          Text("(c)", 0, 0, size=12, weight='bold'),
                      ).move(550, 350),
                     )
final_figure

final_figure.save(f"assets/figures/{FIGURE}/{FIGURE}.svg")
svg2png(url=f"assets/figures/{FIGURE}/{FIGURE}.svg", write_to=f"assets/figures/{FIGURE}/{FIGURE}.png")
