==> 11_FigS9final.python.ipynb <==
from pathlib import Path
from svgutils.compose import *
from svgutils.compose import Figure

FIGURE1_FOLDER = "Figure_5"
FIGURE1 = "Figure_5"
FIGURE = "Figure_S9"
cutoff1, cutoff2, cutoff3 = "0.30", "0.40", "0.50"

scale = 0.65
x1, x2 = 0, 400
y1 = 0
height = 370
letter_size = 18
text_size = 12
final_figure = Figure("800", "1150",
                      Panel(
                          SVG(f"assets/figures/{FIGURE1_FOLDER}/{FIGURE1}C_{cutoff1}.svg").scale(scale).move(x1, y1),
                          Text("A", x1+10, y1+55, size=letter_size, weight='bold'),
                          Text(f"BiG-SCAPE Cutoff: {cutoff1}", x1+10, y1+370, size=text_size, weight='bold'),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE1_FOLDER}/{FIGURE1}D_{cutoff1}.svg").scale(scale).move(x2, y1),
                          Text("B", x2+10, y1+55, size=letter_size, weight='bold'),
                          Text(f"BiG-SCAPE Cutoff: {cutoff1}", x2+10, y1+370, size=text_size, weight='bold'),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}C_{cutoff2}.svg").scale(scale).move(x1, y1 + height),
                          Text("C", x1+10, y1+height+55, size=letter_size, weight='bold'),
                          Text(f"BiG-SCAPE Cutoff: {cutoff2}", x1+10, y1+height+370, size=text_size, weight='bold'),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}D_{cutoff2}.svg").scale(scale).move(x2, y1 + height),
                          Text("D", x2+10, y1+height+55, size=letter_size, weight='bold'),
                          Text(f"BiG-SCAPE Cutoff: {cutoff2}", x2+10, y1+height+370, size=text_size, weight='bold'),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}C_{cutoff3}.svg").scale(scale).move(x1, y1 + (2*height)),
                          Text("E", x1+10, y1+(2*height)+55, size=letter_size, weight='bold'),
                          Text(f"BiG-SCAPE Cutoff: {cutoff3}", x1+10, y1+(2*height)+370, size=text_size, weight='bold'),                      
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}D_{cutoff3}.svg").scale(scale).move(x2, y1 + (2*height)),
                          Text("F", x2+10, y1+(2*height)+55, size=letter_size, weight='bold'),
                          Text(f"BiG-SCAPE Cutoff: {cutoff3}", x2+10, y1+(2*height)+370, size=text_size, weight='bold'),
                      )
                     )
final_figure

texts = [
'assets/figures/Figure_5/Figure_5_text_bigscape_only_0.30.txt',
'assets/figures/Figure_5/Figure_5_text_bigscape_integrated_0.30.txt',
'assets/figures/Figure_S9/Figure_S9_text_bigscape_only_0.40.txt',
'assets/figures/Figure_S9/Figure_S9_text_bigscape_integrated_0.40.txt',
'assets/figures/Figure_S9/Figure_S9_text_bigscape_only_0.50.txt',
'assets/figures/Figure_S9/Figure_S9_text_bigscape_integrated_0.50.txt',
]

annotation = []
for i in texts:
    with open(i, "r") as f:
        text = f.readlines()
        annotation.append("".join(text))

print("".join(annotation))

outfile = Path(f"assets/figures/{FIGURE}/{FIGURE}.svg")
outfile.parent.mkdir(parents=True, exist_ok=True)
final_figure.save(outfile)
