==> 16_Fig6_Erythreapeptin_final.python.ipynb <==
## Merge Visualization

from svgutils.compose import *
from svgutils.compose import Figure
from IPython.display import SVG as disp_SVG
from pathlib import Path

FIGURE = "Figure_6"

#create lines
heights = [30, 40, 50, 60, 70]

for height in heights:
    line = f"""<?xml version="1.0" encoding="UTF-8" standalone="no"?>
    <svg xmlns="http://www.w3.org/2000/svg" width="10" height="{height}">
      <line x1="10" y1="0" x2="10" y2="{height}" style="stroke: black;"/>
    </svg>"""

    outfile = Path(f"assets/figures/Figure_6/doodles/line_{height}.svg")
    outfile.parent.mkdir(parents=True, exist_ok=True)
    with open(outfile, "w") as f:
        f.write(line)

width, height = 10, 270
# create boxes
for width in [133, 31, 70, 51, 24, 190]:
    rectangle = f"""<svg width="{width+50}" height="{height+50}">
    <rect x="20" y="20" width="{width}" height="{height}"
    style="fill:none;stroke:black;stroke-width:0.5;opacity:0.5" />
    </svg>"""

    outfile = Path(f"assets/figures/Figure_6/doodles/rectangle_{width}x{height}x.svg")
    outfile.parent.mkdir(parents=True, exist_ok=True)
    with open(outfile, "w") as f:
            f.write(rectangle)

# precursor peptide

color = "red"
# create boxes
precursor_box_width, precursor_box_height = 190, 320
width, height = precursor_box_width, precursor_box_height
for width in [width]:
    rectangle = f"""<svg width="{width+50}" height="{height+50}">
    <rect x="20" y="20" width="{width}" height="{height}"
    style="fill:none;stroke:{color};stroke-width:1;opacity:1" />
    </svg>"""

    outfile = Path(f"assets/figures/Figure_6/doodles/rectangle_{width}x{height}_{color}.svg")
    outfile.parent.mkdir(parents=True, exist_ok=True)
    with open(outfile, "w") as f:
            f.write(rectangle)

disp_SVG(rectangle)

cutoff = "0.30"

with open(f"assets/figures/{FIGURE}/{FIGURE}d.svg", "r") as f:
    context = f.readlines()

for num, i in enumerate(context):
    if "consensus" in i:
        for coor in context[num-1].split():
            if 'x=' in coor:
                precursor_x = float(coor.strip("x=").strip("'"))

a_x, a_y = 0, 30  
b_x, b_y = 450, 30
c_x, c_y = 0, 450
d_x, d_y = 450, 450
gggenomes_x = c_x + 350

precursor_scale = 1
core_peptide_x, core_peptide_y = d_x+precursor_x-(200*precursor_scale), d_y+(300*precursor_scale)


final_figure = Figure("1200", "1200",
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/a_annotated.svg").scale(0.75).move(a_x, a_y),
                          Text("A", a_x, a_y, size=18, weight='bold'),  
                          SVG(f"assets/figures/{FIGURE}/doodles/rectangle_133x270x.svg").move(a_x+97, a_y+1),
                          SVG(f"assets/figures/{FIGURE}/doodles/rectangle_31x270x.svg").move(a_x+110+120, a_y+1),
                          SVG(f"assets/figures/{FIGURE}/doodles/rectangle_70x270x.svg").move(a_x+110+120+31, a_y+1),
                          SVG(f"assets/figures/{FIGURE}/doodles/rectangle_51x270x.svg").move(a_x+110+120+31+70, a_y+1),
                          SVG(f"assets/figures/{FIGURE}/doodles/rectangle_24x270x.svg").move(a_x+110+120+31+70+51, a_y+1),
                          Text("I", a_x+180, a_y+30, size=10, weight='bold'),  
                          Text("II", a_x+260, a_y+30, size=10, weight='bold'), 
                          Text("III", a_x+305, a_y+30, size=10, weight='bold'), 
                          Text("IV", a_x+370, a_y+30, size=10, weight='bold'), 
                          Text("V", a_x+410, a_y+30, size=10, weight='bold'), 
                      ),                         
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}BD_2_{cutoff}.svg").scale(0.6).move(b_x+0, b_y + 0),
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}BD_1_{cutoff}.svg").scale(0.25).move(b_x+250, b_y - 20),
                          Text("BGC0000513", b_x+200, b_y+170, size=10, weight='bold'),
                          Text("Erythreapeptin", b_x+200, b_y+180, size=10, weight='light'),
                          Text("GCF_150", b_x+140, b_y+100, size=10, weight='bold'),
                          Text("GCF_299", b_x+160, b_y+120, size=10, weight='bold'),
                          Text("GCF_77", b_x+100, b_y+160, size=10, weight='bold'),
                          Text("GCF_64", b_x+80, b_y+260, size=10, weight='bold'),
                          Text("GCF_164", b_x+165, b_y+290, size=10, weight='bold'),
                          Text("GCF_8", b_x+290, b_y+220, size=10, weight='bold'),
                          Text("B", b_x, b_y, size=18, weight='bold'),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}c.svg").scale(0.9).move(c_x, c_y-30),
                          Text("C", c_x, c_y , size=18, weight='bold'),
                          #Text("P6 | NZ_JADDUE010000019.1.region001", c_x+70, c_y+270, size=6.5, color="black"), 
                          Text("GCF_8", gggenomes_x, c_y+60, size=10, color="black", weight='bold'), 
                          SVG(f"assets/figures/{FIGURE}/doodles/line_70.svg").move(gggenomes_x-15, c_y+18),
                          #Text("GCF_77", gggenomes_x+50+30, c_y+130, size=10, color="black", weight='bold'),
                          SVG(f"assets/figures/{FIGURE}/doodles/line_40.svg").move(gggenomes_x+55, c_y+105),
                          Text("GCF_299", gggenomes_x, c_y+162.5, size=10, color="black", weight='bold'),
                          Text("GCF_150", gggenomes_x+35, c_y+180, size=10, color="black", weight='bold'),
                          Text("GCF_164", gggenomes_x+35, c_y+200, size=10, color="black", weight='bold'),
                          Text("GCF_64", gggenomes_x+50, c_y+240, size=10, color="black", weight='bold'),
                          SVG(f"assets/figures/{FIGURE}/doodles/line_60.svg").move(gggenomes_x+30, c_y+210),
                          Text("EryKC", c_x+235, c_y-10, size=8, color="black", weight='light').rotate(-45, c_x+235, c_y-10),
                          Text("EryS", c_x+235+10, c_y-10, size=8, color="black", weight='light').rotate(-45, c_x+235+10, c_y-10),
                          Text("ramA", c_x+235+20, c_y-10, size=8, color="black", weight='normal').rotate(-45, c_x+235+20, c_y-10),
                          Text("ramB", c_x+235+30, c_y-10, size=8, color="black", weight='normal').rotate(-45, c_x+235+30, c_y-10),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}d.svg").scale(precursor_scale).move(d_x, d_y),
                          Text("D", d_x, d_y, size=18, weight='bold'), 
                          Text("GCF_77", gggenomes_x+50+20, c_y+130, size=10, color="black", weight='bold'),
                          SVG(f"assets/figures/{FIGURE}/doodles/rectangle_{precursor_box_width}x{precursor_box_height}_red.svg").move(core_peptide_x, d_y-20),
                          Text("Core Peptide", core_peptide_x, core_peptide_y, size=10, weight='bold'),
                          Text("Leader Peptide", d_x+160, d_y+300, size=10, weight='bold'),
                          Text("Precursor Peptide Alignment", d_x+50, d_y-5, size=12, weight='bold'),
                      ),
                     )
final_figure

outfile = Path(f"assets/figures/{FIGURE}/{FIGURE}_{cutoff}_MSA.svg")
outfile.parent.mkdir(parents=True, exist_ok=True)
final_figure.save(outfile)

[](assets/figures/Figure_6/b.svg)
