==> 19_Fig7_Mycofactocin_final_figure.python.ipynb <==
## Merge Visualization

from svgutils.compose import *
from svgutils.compose import Figure
from pathlib import Path

FIGURE = "Figure_7"

cutoff = "0.30"
final_figure = Figure("900", "800",
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}B_{cutoff}.svg").scale(0.5).move(0, 30),
                          #Text("A", 30, 30, size=18, weight='bold'),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}D_{cutoff}.svg").scale(0.2).move(200, 0),
                          Text("A", 0, 30, size=18, weight='bold'),
                      ),
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/b.svg").scale(0.8).move(330, 0),
                          Text("B", 350, 30, size=18, weight='bold'),
                      ),                      
                      Panel(
                          SVG(f"assets/figures/{FIGURE}/{FIGURE}c.svg").scale(0.75).move(350, 350),
                          Text("D", 350, 350, size=18, weight='bold'),                        
                      ),                      
                      Panel(
                          #SVG(f"assets/figures/Figure_7/Mycofactocin_Biosynthesis_Annotated.svg").scale(0.75).move(350, 350),
                          Text("C", 0, 350, size=18, weight='bold'),                        
                      ),
                     )
outfile = Path(f"assets/figures/{FIGURE}/{FIGURE}_{cutoff}_new.svg")
outfile.parent.mkdir(parents=True, exist_ok=True)
final_figure.save(outfile)
final_figure
