==> 99_Zenodo.ipynb <==
## Upload Dataset to Zenodo
https://developers.zenodo.org/?python#quickstart-upload

import requests
r = requests.get("https://zenodo.org/api/deposit/depositions")
r.status_code
# 401
r.json()

ACCESS_TOKEN = CHANGE_THIS_FIELD

r = requests.get('https://zenodo.org/api/deposit/depositions',
                  params={'access_token': ACCESS_TOKEN})
r.status_code
# 200
r.json()
# []

# pick your project, see docs to create project from scratch
json = r.json()[0]
json

headers = {"Content-Type": "application/json"}
params = {'access_token': ACCESS_TOKEN}
r = requests.post('https://zenodo.org/api/deposit/depositions',
                   params=params,
                   json=json)
r.status_code
# 201
r.json()

bucket_url = r.json()["links"]["bucket"]
bucket_url

filename = "saccharopolyspora_dataset.zip"
path = "../%s" % filename

''' 
The target URL is a combination of the bucket link with the desired filename
seperated by a slash.
'''
with open(path, "rb") as fp:
    r = requests.put(
        "%s/%s" % (bucket_url, filename),
        data=fp,
        params=params,
    )
r.json()
